+++ main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"mcpsync/cmd"
-	_ "mcpsync/internal/agent"
+	"github.com/anush-data-portfolio/MCPSync/cmd"
+	_ "github.com/anush-data-portfolio/MCPSync/internal/agent"
 )
 
 func main() {